requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
sentence-transformers==2.2.2
numpy==1.26.4
scikit-learn==1.4.2